Update slam3.py np.float

+++ slam/slam3.py
@@ -535,7 +535,7 @@ class Slam3(object):
 
         elif sample_weight_scheme is "bool":
             # 0|1 scheme for training flux (recommended)
-            sample_weight = self.tr_mask.astype(np.float)
+            sample_weight = self.tr_mask.astype(np.float32)
             ind_all_bad = np.sum(sample_weight, axis=0) < 1.
             for i_pix in np.arange(sample_weight.shape[1]):
                 if ind_all_bad[i_pix]:
@@ -721,7 +721,7 @@ class Slam3(object):
 
         # test_ivar default
         if test_ivar is None:
-            test_ivar = np.ones_like(test_flux, dtype=np.float)
+            test_ivar = np.ones_like(test_flux, dtype=np.float32)
 
         # test_ivar normalization --> unnecessary
         #test_ivar /= np.nansum(test_ivar[mask])
@@ -1031,7 +1031,7 @@ class Slam3(object):
         # test_ivar must be set here!
         if test_ivar is None:
             # test_ivar=None, directly set test_ivar
-            test_ivar = np.ones_like(test_flux, dtype=np.float)
+            test_ivar = np.ones_like(test_flux, dtype=np.float32)
         # test_ivar is not None
         elif ivar_scaler is not None:
             # do scaling for test_ivar
@@ -1194,7 +1194,7 @@ class Slam3(object):
         # test_ivar must be set here!
         if test_ivar is None:
             # test_ivar=None, directly set test_ivar
-            test_ivar = np.ones_like(test_flux, dtype=np.float)
+            test_ivar = np.ones_like(test_flux, dtype=np.float32)
         # test_ivar is not None
         elif ivar_scaler is not None:
             # do scaling for test_ivar
@@ -1576,7 +1576,7 @@ class Slam3(object):
                                (self.tr_flux > 0.) *
                                np.isfinite(self.tr_ivar) *
                                np.isfinite(self.tr_flux))
-            sample_weight = ind_good_pixels.astype(np.float)
+            sample_weight = ind_good_pixels.astype(np.float32)
             r = Parallel(n_jobs=n_jobs, verbose=verbose)(delayed(nmse)(
                 self.sms[i], self.tr_labels_scaled, self.tr_flux_scaled[:, i],
                 sample_weight[:, i]) for i in range(self.n_pix))
